requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
scipy==1.11.4
matplotlib==3.8.4
scikit-learn==1.4.2
pandas==2.2.2
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
Flask==3.0.3
streamlit==1.35.0
minio==7.2.7
boto3==1.34.131
